refactor: organize frontend types by domain
